Time: 2 ms (74.27%), Space: 26.7 MB (15.31%) - LeetHub

+++ 0680-valid-palindrome-ii/0680-valid-palindrome-ii.cpp
@@ -1,6 +1,6 @@
 class Solution {
 
-bool checkPalindrome(string s, long long int i,long long int j){
+bool checkPalindrome(string s, int i,int j){
     while(i<=j){
     if(s[i] != s[j]){
         return false;
@@ -13,8 +13,8 @@ bool checkPalindrome(string s, long long int i,long long int j){
 
 public:
     bool validPalindrome(string s) {
-        long long int i=0;
-        long long int j= s.length()-1;
+        int i=0;
+        int j= s.length()-1;
         while(i<=j){
             if(s[i]!= s[j]){
                 return checkPalindrome(s, i+1, j) || checkPalindrome(s, i, j-1);
